Split workspace generation iteration runner
